+++ src/attaque/Tranchant.java
@@ -2,7 +2,7 @@ package attaque;
 
 public abstract class Tranchant extends Pouvoir{
 
-	public Tranchant(int pointDeDegat, String nom, int nbUtilisationPouvoir) {
+	public Tranchant(int pointDeDegat, String nom, double nbUtilisationPouvoir) {
 		super(pointDeDegat, nom, nbUtilisationPouvoir);
 	}
 }
